Fix bug in crystfel.py
There are some fundamental issues with the way that Panel and
PanelList objects deal with data and mask arrays.  The real issue
is that there are duplicates of the same data, and they get out
of sync.  Need to completely re-think this.  Generally, there
should not be data arrays hard-coded into these classes; that
idea was wrong from the beginning since there are too many types
of data that one might want to work with: intensities, masks of
various types, mean intensities, amplitudes, etc.

A better approach would be to just create the convenience methods
for slicing arrays.  E.g. a method for creating a view of a
PanelList array that is specific to a particular Panel.