Remove data arrays from detector.Panel class
This changes the way we work with detector Panels.  It simplifies
things a lot, since the management of data arrays was becoming a
headache.  Now we think of Panels as simply a container for
detector geometry.  Note that PanelList should be considered
the fundamental class for diffraction data; you should work with
PanelLists rather than panels.