Don't show mpl displays during testing by default.
Sorry Derek - I like having gui tests but it's a bit
of a nuisance to have the window pop up by default.
I run tests often, and now I need to find the window
and manually close it with the mouse every time...
Best if this is optional - I just don't know how to
set that up.